+++ src/app/containers/PageHandlers/withData.jsx
@@ -5,42 +5,48 @@ import ErrorMain from '../ErrorMain';
 import nodeLogger from '../../helpers/logger.node';
 import deepGet from '../../helpers/json/deepGet';
 import { ServiceContext } from '../../contexts/ServiceContext';
-import { getArticleHomeService } from '../PageViewAnalytics/labelHelpers/article';
+import getPassportHome from '../../lib/utilities/getPassportHome';
 
 const logger = nodeLogger(__filename);
 
-// checks for data, status & home keys, setting default status if not found
+// checks for data, status, setting default status if not found
 const constructRenderObject = data => ({
   status: deepGet(['status'], data) || 500,
   pageData: deepGet(['pageData'], data),
-  home: deepGet(['pageData', 'metadata', 'passport', 'home'], data),
+  type: deepGet(['pageData', 'metadata', 'type'], data),
 });
 
 // checks for pageData, 200 status and if home service from article data fits the service locale
 const shouldRender = data => {
-  const { status, pageData } = constructRenderObject(data);
+  const { status, pageData, type } = constructRenderObject(data);
   // eslint-disable-next-line react-hooks/rules-of-hooks
   const { service } = useContext(ServiceContext);
 
   const hasDataAnd200Status = pageData && status === 200;
-  const homeService = pageData && getArticleHomeService(pageData);
+  const passportHome = pageData && getPassportHome(pageData);
 
-  const isValidHome = pageData && homeService === service;
+  const isValidHome = pageData && passportHome === service;
 
   return {
     hasDataAnd200Status,
     isValidHome,
     status: hasDataAnd200Status && !isValidHome ? 404 : status,
     pageData,
+    type,
   };
 };
 
 const WithData = Component => {
   const DataContainer = ({ data, ...props }) => {
-    const { hasDataAnd200Status, isValidHome, status, pageData } = shouldRender(
-      data,
-    );
-    if (hasDataAnd200Status && isValidHome) {
+    const {
+      hasDataAnd200Status,
+      isValidHome,
+      status,
+      pageData,
+      type,
+    } = shouldRender(data);
+
+    if ((hasDataAnd200Status && isValidHome) || type === 'IDX') {
       try {
         return <Component pageData={pageData} {...props} />;
       } catch (err) {
